WellIndexCalculator is now a separate library
