+++ FieldOpt/CMakeLists.txt
@@ -35,6 +35,7 @@ enable_testing()
 add_subdirectory(Utilities)
 add_subdirectory(Settings)
 add_subdirectory(ERTWrapper)
+add_subdirectory(WellIndexCalculator)
 add_subdirectory(Reservoir)
 add_subdirectory(Model)
 add_subdirectory(Optimization)
