+++ FieldOpt/Model/wells/wellbore/wellspline.h
@@ -4,7 +4,7 @@
 #include "trajectory.h"
 #include "Reservoir/grid/eclgrid.h"
 #include "Model/wells/wellbore/wellblock.h"
-#include "Reservoir/well_index_calculation/wellindexcalculator.h"
+#include "WellIndexCalculator/wellindexcalculator.h"
 #include <QList>
 
 namespace Model {
