+++ FieldOpt/WellIndexCalculator/intersected_cell.cpp
