+++ FieldOpt/WellIndexCalculator/intersected_cell.h
