+++ FieldOpt/WellIndexCalculator/wellindexcalculator.cpp
