+++ FieldOpt/WellIndexCalculator/wellindexcalculator.h
